Switch from SQS to SNS
Now when a new comment is posted to post-comments lambda, it publishes the comment
to an SNS topic instead of an SQS queue. I made the change because the
create-pull-request lambda was constantly polling the SQS endpoint to see if it
should be run. This was eating up almost all of my free SQS quota.

I also cleaned up the code a little bit based on more experience working with lambda
and nodejs.